+++ kafka_influxdb/encoder/errors.py
@@ -1,0 +1,10 @@
+"""
+Custom errors of encoders
+"""
+
+class EncoderError(Exception):
+    """
+    Error during encoding
+    """
+    pass
+