+++ kafka_influxdb/tests/test_worker.py
@@ -3,13 +3,19 @@ from mock import Mock
 import random
 from kafka_influxdb.worker import Worker
 from kafka_influxdb.encoder import echo_encoder
+from kafka_influxdb.tests.helpers.timeout import timeout
+from kafka.common import ConnectionError
 
 
 class Config:
+    """
+    Dummy config with minimum settings to pass the tests
+    """
     def __init__(self, buffer_size):
         self.buffer_size = buffer_size
         self.kafka_topic = "test"
         self.influxdb_dbname = "mydb"
+        self.statistics = False
 
 
 class DummyReader(object):
@@ -23,6 +29,8 @@ class DummyReader(object):
     def read(self):
         for i in range(self.num_messages):
             yield random.choice(self.messages)
+        # Simulate keyboard interrupt by user to stop consuming
+        raise KeyboardInterrupt
 
 
 class FlakyReader(object):
@@ -46,42 +54,37 @@ class FlakyReader(object):
             # Yield the second half of messages
             for i in range(int(self.num_messages/2)):
                 yield self.message
+        # Simulate keyboard interrupt by user to stop consuming
+        raise KeyboardInterrupt
 
-class DummyWriter(object):
+
+class TestWorker(unittest.TestCase):
     """
-    A fake writer that does nothing with the input data
+    Tests for message worker.
     """
-    def __init__(self):
-        pass
-
-    def write(self):
-        pass
-
-
-class TestKafkaInfluxDB(unittest.TestCase):
     def setUp(self):
-        self.config = Config(100)
+        self.config = Config(10)
         self.encoder = echo_encoder.Encoder()
-        self.writer = DummyWriter()
         self.writer = Mock()
         self.writer.write.return_value = True
 
-    def test_buffering(self):
-        self.reader = DummyReader(["foo"], self.config.buffer_size - 1)
-        self.client = Worker(self.reader, self.encoder, self.writer, self.config)
-        self.client.consume()
-        self.assertFalse(self.writer.write.called)
-
     def test_flush(self):
-        self.reader = DummyReader(["bar"], self.config.buffer_size)
-        self.client = Worker(self.reader, self.encoder, self.writer, self.config)
-        self.client.consume()
+        """
+        Messages should be flushed to the writer when the buffer is full.
+        """
+        reader = DummyReader(["bar"], self.config.buffer_size)
+        client = Worker(reader, self.encoder, self.writer, self.config)
+        client.consume()
         self.assertTrue(self.writer.write.called)
         self.writer.write.assert_called_once_with(["bar"] * self.config.buffer_size)
 
+    @timeout(0.1)
     def test_reconnect(self):
-        self.reader = FlakyReader("baz", self.config.buffer_size)
-        self.client = Worker(self.reader, self.encoder, self.writer, self.config)
-        self.client.consume()
+        """
+        The worker should reconnect when the connection is interrupted.
+        """
+        reader = FlakyReader("baz", self.config.buffer_size)
+        client = Worker(reader, self.encoder, self.writer, self.config)
+        client.consume()
         self.assertTrue(self.writer.write.called)
         self.writer.write.assert_called_once_with(["baz"] * self.config.buffer_size)