+++ kafka_influxdb/writer/influxdb_writer.py
@@ -52,7 +52,6 @@ class InfluxDBWriter(object):
 
         logging.info("Connecting to InfluxDB at %s:%s (SSL: %r, UDP: %r)", host, port, use_ssl, use_udp)
         self.client = self.create_client()
-        logging.info("Creating database %s if not exists", dbname)
 
     def create_client(self):
         """
@@ -92,6 +91,7 @@ class InfluxDBWriter(object):
             params = self.params
 
         try:
+            logging.debug("Writing message: %s", msg)
             self.client.request(url='write',
                                 method='POST',
                                 params=params,